Install everything in /root for the root user.
We are in a development VM environment, so it makes sense to just use
the root user for most tasks for now.

Interested parties can check the various playbooks and ensure commands
are run as the right user and end up with the right file permissions.
See http://docs.ansible.com/playbooks_intro.html#hosts-and-users